tustus_2.7.4b/requirements.txt
python-telegram-bot[job-queue]==21.6
requests
beautifulsoup4
lxml
